requirements.txt
dj-database-url==0.4.1
Django==1.10.3
djangorestframework==3.5.3
gunicorn==19.6.0
virtualenv==15.1.0
whitenoise==3.2.2
